requirements.txt
mysql_connector_repackaged==0.3.1
numpy==2.1.2
pandas==1.4.0
Requests==2.32.3
